update: 마이페이지 애니메이션
