+++ src/app/my-page/components/mypage-profile.tsx
@@ -80,7 +80,7 @@ const MypageProfile = ({ userId }: { userId: string }) => {
   if (isPending) return <div>쿠키 주문하신 분?</div>;
 
   return (
-    <div className="relative w-full rounded-2xl border border-primary-50 bg-white py-5 md:flex md:h-[300px] md:w-[600px] md:items-center md:justify-center">
+    <>
       <div className="flex flex-col items-center gap-2">
         <EmotionImage src={EMOTION_COOKIE_IMAGE_URL.EXCITED} size="l" />
         <div className="flex h-10 items-center justify-center gap-2">
@@ -134,7 +134,7 @@ const MypageProfile = ({ userId }: { userId: string }) => {
           <strong className="text-sm text-grey-3">나의 프로필</strong>
         )}
       </div>
-    </div>
+    </>
   );
 };
